+++ src/util/size.rs
@@ -8,9 +8,9 @@ pub enum ParseSizeError {
 
 #[derive(Clone, Debug)]
 pub enum Size {
-    Percent(usize),
-    Pixel(usize),
-    Raw(usize),
+    Percent(f32),
+    Pixel(f32),
+    Raw(f32),
 }
 
 impl ToString for Size {
@@ -28,7 +28,7 @@ impl Size {
         let re = Regex::new(r"^(\d+)px$").unwrap();
         re.captures(input)
             .and_then(|list| list.get(1))
-            .and_then(|first| first.as_str().parse::<usize>().ok())
+            .and_then(|first| first.as_str().parse::<f32>().ok())
             .and_then(|value| Some(Size::Pixel(value)))
     }
 
@@ -36,18 +36,18 @@ impl Size {
         let re = Regex::new(r"^(\d+)%$").unwrap();
         re.captures(input)
             .and_then(|list| list.get(1))
-            .and_then(|first| first.as_str().parse::<usize>().ok())
+            .and_then(|first| first.as_str().parse::<f32>().ok())
             .and_then(|value| Some(Size::Percent(value)))
     }
 
     fn parse_raw(input: &str) -> Option<Size> {
         input
-            .parse::<usize>()
+            .parse::<f32>()
             .ok()
             .and_then(|value| Some(Size::Raw(value)))
     }
 
-    pub fn value(&self) -> usize {
+    pub fn value(&self) -> f32 {
         match self {
             Size::Percent(value) => value.clone(),
             Size::Pixel(value) => value.clone(),
